Also move {ssh,sftp}-user commands
