+++ package.json
@@ -167,15 +167,6 @@
       "project": {
         "description": "Manage your projects, and also any kinds of user memberships concerning these projects.",
         "subtopics": {
-          "backup": {
-            "description": "Manage backups of your projects"
-          },
-          "backupschedule": {
-            "description": "Manage backup schedules of your projects"
-          },
-          "cronjob": {
-            "description": "Manage cronjobs of your projects"
-          },
           "filesystem": {
             "description": "Interact with the filesystem of your project"
           },
@@ -184,17 +175,17 @@
           },
           "membership": {
             "description": "Control who gets to work on your projects, and who doesn't"
-          },
-          "sftp-user": {
-            "description": "Manage SFTP users of your projects"
-          },
-          "ssh-user": {
-            "description": "Manage SSH users of your projects"
           }
         }
       },
       "server": {
         "description": "Manage your servers"
+      },
+      "sftp-user": {
+        "description": "Manage SFTP users of your projects"
+      },
+      "ssh-user": {
+        "description": "Manage SSH users of your projects"
       },
       "user": {
         "description": "Manage your own user account",
